requirements.txt
flask
pandas
numpy
tensorflow
yfinance
newsapi-python
vaderSentiment
joblib
openpyxl
plotly
requests